+++ src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/nodes/interpreter/InterpreterV3PlusClosuresNode.java
@@ -723,7 +723,7 @@ public final class InterpreterV3PlusClosuresNode extends AbstractInterpreterNode
                     case BC.BYTECODE_PRIM_ADD: {
                         final Object arg = pop(frame, --sp);
                         final Object receiver = pop(frame, --sp);
-                        final byte profile = getProfile(currentPC);
+                        final short profile = getProfile(currentPC);
                         final Object result;
                         if (receiver instanceof final Long lhs && arg instanceof final Long rhs) {
                             enter(currentPC, profile, BRANCH2);
@@ -751,7 +751,7 @@ public final class InterpreterV3PlusClosuresNode extends AbstractInterpreterNode
                     case BC.BYTECODE_PRIM_SUBTRACT: {
                         final Object arg = pop(frame, --sp);
                         final Object receiver = pop(frame, --sp);
-                        final byte profile = getProfile(currentPC);
+                        final short profile = getProfile(currentPC);
                         final Object result;
                         if (receiver instanceof final Long lhs && arg instanceof final Long rhs) {
                             enter(currentPC, profile, BRANCH2);
@@ -779,7 +779,7 @@ public final class InterpreterV3PlusClosuresNode extends AbstractInterpreterNode
                     case BC.BYTECODE_PRIM_LESS_THAN: {
                         final Object arg = pop(frame, --sp);
                         final Object receiver = pop(frame, --sp);
-                        final byte profile = getProfile(currentPC);
+                        final short profile = getProfile(currentPC);
                         final Object result;
                         if (receiver instanceof final Long lhs && arg instanceof final Long rhs) {
                             enter(currentPC, profile, BRANCH2);
@@ -799,7 +799,7 @@ public final class InterpreterV3PlusClosuresNode extends AbstractInterpreterNode
                     case BC.BYTECODE_PRIM_GREATER_THAN: {
                         final Object arg = pop(frame, --sp);
                         final Object receiver = pop(frame, --sp);
-                        final byte profile = getProfile(currentPC);
+                        final short profile = getProfile(currentPC);
                         final Object result;
                         if (receiver instanceof final Long lhs && arg instanceof final Long rhs) {
                             enter(currentPC, profile, BRANCH2);
@@ -819,7 +819,7 @@ public final class InterpreterV3PlusClosuresNode extends AbstractInterpreterNode
                     case BC.BYTECODE_PRIM_LESS_OR_EQUAL: {
                         final Object arg = pop(frame, --sp);
                         final Object receiver = pop(frame, --sp);
-                        final byte profile = getProfile(currentPC);
+                        final short profile = getProfile(currentPC);
                         final Object result;
                         if (receiver instanceof final Long lhs && arg instanceof final Long rhs) {
                             enter(currentPC, profile, BRANCH2);
@@ -839,7 +839,7 @@ public final class InterpreterV3PlusClosuresNode extends AbstractInterpreterNode
                     case BC.BYTECODE_PRIM_GREATER_OR_EQUAL: {
                         final Object arg = pop(frame, --sp);
                         final Object receiver = pop(frame, --sp);
-                        final byte profile = getProfile(currentPC);
+                        final short profile = getProfile(currentPC);
                         final Object result;
                         if (receiver instanceof final Long lhs && arg instanceof final Long rhs) {
                             enter(currentPC, profile, BRANCH2);
@@ -859,7 +859,7 @@ public final class InterpreterV3PlusClosuresNode extends AbstractInterpreterNode
                     case BC.BYTECODE_PRIM_EQUAL: {
                         final Object arg = pop(frame, --sp);
                         final Object receiver = pop(frame, --sp);
-                        final byte profile = getProfile(currentPC);
+                        final short profile = getProfile(currentPC);
                         final Object result;
                         if (receiver instanceof final Long lhs && arg instanceof final Long rhs) {
                             enter(currentPC, profile, BRANCH2);
@@ -879,7 +879,7 @@ public final class InterpreterV3PlusClosuresNode extends AbstractInterpreterNode
                     case BC.BYTECODE_PRIM_NOT_EQUAL: {
                         final Object arg = pop(frame, --sp);
                         final Object receiver = pop(frame, --sp);
-                        final byte profile = getProfile(currentPC);
+                        final short profile = getProfile(currentPC);
                         final Object result;
                         if (receiver instanceof final Long lhs && arg instanceof final Long rhs) {
                             enter(currentPC, profile, BRANCH2);
@@ -899,7 +899,7 @@ public final class InterpreterV3PlusClosuresNode extends AbstractInterpreterNode
                     case BC.BYTECODE_PRIM_BIT_AND: {
                         final Object arg = pop(frame, --sp);
                         final Object receiver = pop(frame, --sp);
-                        final byte profile = getProfile(currentPC);
+                        final short profile = getProfile(currentPC);
                         final Object result;
                         if (receiver instanceof final Long lhs && arg instanceof final Long rhs) {
                             enter(currentPC, profile, BRANCH2);
@@ -916,7 +916,7 @@ public final class InterpreterV3PlusClosuresNode extends AbstractInterpreterNode
                     case BC.BYTECODE_PRIM_BIT_OR: {
                         final Object arg = pop(frame, --sp);
                         final Object receiver = pop(frame, --sp);
-                        final byte profile = getProfile(currentPC);
+                        final short profile = getProfile(currentPC);
                         final Object result;
                         if (receiver instanceof final Long lhs && arg instanceof final Long rhs) {
                             enter(currentPC, profile, BRANCH2);
